+++ packages/plugin-gallery/web/src/getAddButtons.tsx
@@ -35,9 +35,12 @@ const handleNativeFileChange =
   (editorCommands, uploadService, uploader, galleryPluginService) => (files: File[]) => {
     const nodeId = editorCommands.insertBlockWithBlankLines(GALLERY_TYPE, defaultData);
     files.forEach(file => {
-      setTimeout(
-        () => uploadService.uploadFile(file, nodeId, uploader, GALLERY_TYPE, galleryPluginService),
-        0
+      uploadService.uploadFile(
+        new File([file], file.name, { type: file.type }),
+        nodeId,
+        uploader,
+        GALLERY_TYPE,
+        galleryPluginService
       );
     });
   };